Extract travel highlight section
